+++ frontend/eventsync_front/src/views/ThemeRoomViews/ListTrView.vue
@@ -4,7 +4,7 @@
     </v-app>
 </template>
 <script setup lang="ts">
-import ListComponentTR from '../../components/ThemeRoom/ListComponentTR.vue'
+import ListComponentTR from '../../components/ThemeRoom/ListRooms.vue'
 // import NavBar from '../../components/NavBar.vue'
 // import FooterVue from '../../components/Footer.vue'
 </script>
